fixed delete_library crashing bug

+++ Code/Core/core.cpp
@@ -573,17 +573,18 @@ void Core::delete_entire_library(const std::string& extra_folder_to_delete) {
 
     // TODO: PUT EXTRA STEPS IN PLACE TO ENSURE SAFETY WHEN CALLING !!!
 
-    // delete entire music files folder
-    FileManager::delete_folder(MUSIC_FOLDER);
-
-    // delete entire images folder
-    FileManager::delete_folder(IMAGES_FOLDER);
-
-    // delete entire database file
+    // delete & remake entire database file
     FileManager::delete_file("Code/Core/Database/music_manager.db"); // must match path in database_manager.h
+    database = DatabaseManager{}; // remake database file
 
-    // remake database file
-    database = DatabaseManager{};
+    // delete entire music files folder (if exists)
+    if (FileManager::exists(MUSIC_FOLDER)) {
+        FileManager::delete_folder(MUSIC_FOLDER);
+    }
+    // delete entire images folder (if exists)
+    if (FileManager::exists(IMAGES_FOLDER)) {
+        FileManager::delete_folder(IMAGES_FOLDER);
+    }
     
     // delete other folder (if passed in)
     if (!extra_folder_to_delete.empty()) { // check if there is extra_folder_to_delete
